Generate impl functions with syn::ImplItem
This commit replaces the generation of impl functions for builder structs using quote! to produce a Tokens stream, with an implementation that generates a vector of ImplItem. In doing so, the ImplItem's can be deduplicated by identifier, to allow common functions to be combined with builder specific functions, when generating the final impl source.