pattern match deserializing response
This commit pattern matches on the attempt to deserialize the response to JSON. Right now, any error is ignored, but this should be surfaced in future.

An alternative would be to deserialize the response as part of a fn on ElasticsearchResponse impl, allowing the user to handle the Result<T> however they see fit.