Include required url parts as new fn arguments
This commit updates code generation to include required url parts common across all API url variants as arguments to the builder new functions, and the name functions on namespace clients that return new instances of builder structs.

The parameters for index, ty (if present) and id should be ordered in a future commit.