Start of url path pattern matching generation
This commit is a Work in Progress on the start of pattern matching to determine and build the correct url for the API, based on the parameters provided. The client lib will currently not compile as the body field is not implemented. The body field requires a refactoring of builder structs and the methods that generate them; APIs that accept a body will require a generic struct where the generic parameter is the body type, defined as the Serialize trait