Allow request body to be supplied for those API that support it
This commit adds the ability to set a request body on those APIs that support sending a body. This requires some branching within code generation to support generic types, where the generic type parameter is the type of body to send, which has a trait bound for serde's Serialize trait. Code generation should be refactored in future to clean up duplication and simplify.

Add a couple of integration tests for search API to assert that sending a request with and without a body work correctly, and that a response can be deserialized.