First attempt at path generation
This commit is a first attempt at generating source code for building url paths,
based on arguments provided by the user. Having individual methods for
"optional" URL parts and pattern matching on tuples in this fashion is not an
elegant way to handle url parts because each collection of parts need to be
handled together. For example, take the Search API which has the following paths

_/search
_/{index}/_search
_/{index}/{type}/_search

Ignoring the fact that the last path is deprecated and will be removed in future,
the optional {type} URL part should not be provided independently of also providing {index} i.e.
{index} is no longer optional when {type} is provided. In this particular API, we can
pass "_all" for {index} when only {type} is provided, but this is not true of all APIs.
Take for example, Get Application Privileges API which has the following paths

/_security/privilege
/_security/privilege/{application}
/_security/privilege/{application}/{privilege}

If {privilege} is provided without also providing {application}, then this is an error,
as there is no wildcard for like "_all" for {application}. Similar issues exist for
Get Repository Snapshot API and others.

This leads to the best way of handling URL parts by modelling them as a
algebraic data types / discriminated union, which is an enum in Rust. I would prefer
not to expose these enums to the end consumer, if possible, and use them internally on
the builder structs only to store URL parts and pattern match to build the URL.

A follow up commit will make the change to enums, but checking this in for now
so that there is a working demonstrable client for the Rust Guild :) There is a
large hack in ./api_generator/code_gen/mod.rs to build exhaustive match arms.