Update REST spec to 7.4.1 and generate url part enums
This commit updates the REST API spec to v7.4.1, updating the structs used to model
the API endpoints.

With the change to group URL parts with each path, this commit also generates enums
for each of the URL parts, replacing the pattern matching on tuples of optional
parts with pattern matching on enums.

Builder constructor functions now accept the associated enum as a ctor argument.
Where an API contains only a single enum value that accepts no URL parts, this
ctor argument is omitted as it is superfluous.

Closes #11
Closes #8