Make associated body functions generic (#18)
This commit makes body functions generic for those APIs that support sending a body, and removes the generic parameter on the associated API method on the root or namespaced_client. This makes it simpler to chain function calls, often with type inference being able to correctly infer the body type T from the argument. It also allows a builder to be cloned and for a different type of body to be set on the clone.

Closes #9