Implement send functions as async
This commit updates the associated send function for each builder struct to be async.
As part of this change, the Sender trait has been removed as it is not currently
possible to have async trait implementations.

Moved integration tests to tests directory, in line with Rust idioms.

The reqwest's crate support for async/await is currently in an alpha release
package, which has a dependency on a crate that uses a feature requiring nightly Rust to
compile. On my machine (Windows), I am running with

> rustup show

active toolchain
----------------

nightly-x86_64-pc-windows-msvc (default)
rustc 1.41.0-nightly (a44774c3a 2019-11-25)

Closes #10