Add Body impls for strings, bytes and slices
This commit adds Body trait impls for String, string slices, byte vectors and slices.
These allow a consumer to use the general purpose send() fn on Elasticsearch to
send request bodies not accounted for by the JsonBody and NdBody implementations.

Update Bytes crate to ^0.5. reqwest dependency is compiled against Bytes ^0.4, so
Bytes From trait impl in reqwest is not compatible for now, requiring Bytes to be
converted to a vec[u8].