Add Body trait to model different API body implementations (#32)
This commit adds a Body trait to allow modelling of different API body implementations.
All APIs accept JSON bodies, except bulk, msearch, msearchtemplate and ML post data,
which accept newline delimited JSON (ndjson).

Provide implementations of Body for 

- json 
- ndjson 
- String and string slices
- Vec<u8> and slices

Serde library is not able to serialize to ndjson because an impl cannot generalize for all 
Serialize impls. To be able to serialize ndjson requires access to the buffer, allowing a Body 
impl to control how it is written to the buffer.

String, string slices, byte vectors and slice Body impls allow a consumer to use the general purpose send() fn on Elasticsearch to send request bodies not accounted for by the 
JsonBody and NdBody implementations.

Use Bytes crate ^0.5. The Reqwest dependency is compiled against Bytes ^0.4, so
Bytes From trait impl in reqwest is not compatible for now, requiring Bytes to be
converted to a vec[u8]. This should be addressed once Reqwest is updated to Bytes ^0.5.

Closes #31