Fix Time enum variants
Relates: https://github.com/elastic/elasticsearch/pull/49368

This commit fixes a bug in the rest specs with the Time enum. The values contain
unit names in parentheses in the spec which results in incorrect enum variants
and serde rename values being generated.

A PR has been submitted to fix the rest spec, which will align with the fix applied in this commit