Refactor errors
This commit is a start of refactoring of errors within the client.
Remove From impl for Reqwest's Response type to Error
as it isn't useful at this stage; a problem with a response is
better handled inside of transport's async send() fn, where an Err could
potentially be returned for a response deemed to be an error.

For now, all HTTP status code responses from Elasticsearch return
an Ok() result. Opened #28 to discuss.