Make the top-level error type opaque (#41)
This commit refactors the top-level Error enum so that it's an opaque struct instead of an open enum. 

This allows changes to the internals of the error without breaking callers relying on the enum having a particular shape. As an example change, we might want to replace the Lib(String) variant with something strongly typed one day. It also means a Backtrace could be introduced, because it'll have a natural place to live within the struct.