Connection url ends in trailing forward slash
This commit fixes a bug in relation to the usage of url::join(). If A url
passed to Connection does not end in a trailing forward slash, a new url
will be constructed from the passed url with a trailing forward slash. When
joining a Connection url with an API path, the leading forward slash of the
path is trimmed, so that the path is appended to the url rather than replacing
any existing path segments.

Fixes #48