BulkOperation builder implementations (#56)
- BulkOperation builder implementations

This commit introduces a collection of builders to help with the bulk API. The bulk
API accepts a collection of operations in newline delimited JSON format, where
each operation is represented as a header, followed by an optional source for the
operation. The builders provide functions to build each of the supported operations,
exposing the metadata options available for the operation. Operations can be converted
to a common Operation type which implements the Body trait.

- Add BulkOperations

  helps in building a collection of operations that vary over
types used to represent source documents.

- Add source filtering to update operation, and introduce
SourceFilter enum to handle the variants.

  Add From<T> implementations for SourceFilter to
ease usage.
 
Closes #43