Attribute source and tests requiring features
This commit attributes source code and tests
that require featues of the elasticsearch package to be
enabled. Currently, this is mainly related to TLS.

Introduce ClientCertificate enum to differentiate between
PKCS#12 archives and PEM certs, which are available
only when native-tls and rustls-tls features are enabled,
respectively.

ClientCertificate enum takes bytes representing
a certificate as opposed to a reqwest Identity because
Identity does not implement Clone, and cannot be dereferenced
from a Credentials::Certificate variant when building a Transport
because other variants are passed to the built Transport.