Add markdown examples into generated client
This commit adds Markdown documentation into the client,
providing a way to document generated APIs.

When generating APIs, files will be looked for in
the elasticsearch/docs directory corresponding to a naming
convention. If found, the contents are included in the generated
doc attribute.

Because the markdown docs are included in the client, examples
are tested when `cargo test` runs.

Closes #66
Closes #67