Allow elasticsearch to work with rust stable (#89)
This commit allows elasticsearch to work with
rust stable. The only feature requiring rust nightly
is external_doc feature to able to to test the README.md
docs. This is now conditionally compiled
when building with rust nightly, using rustc_version
in a build.rs build script to set a cfg flag for the channel
used.

api_generator requires rust nightly still. The default toolchain
to use for each package is in a rust-toolchain file in
the root of each package directory.

Closes #87