Handle url serialization of enum collections
This commit updates serialize_coll_qs to handle serializing
a collection of types that implement Serialize to a comma separated
query string value. This is only needed
for enums and strings, but (mis)uses serde_json serialize to
achieve this, since serde_urlencoded does not support such a format.
We could implement a Serializer for this down the line.