Encode values passed to Parts enums
This commit updates the api_generator to emit code
to percent encode values passed to Parts enums
when calling the url() function. A value passed to a
Part may be one that, if left unencoded, would
be interpreted differently by the HTTP client e.g.
a # (start of Fragment identifier) or ? (start of Query identifier).