Add yaml test runner (#98)
This commit adds a new package to the repository,
yaml_test_runner, that

1. download yaml tests from Elasticsearch repository for a given branch
2. uses yaml test files in conjunction with rest specs to generate
   integration tests for each test defined in yaml.
3. runs tests

Having a yaml test runner for the client brings it in
line with the other official Elasticsearch clients.

Tests are generated in a tests directory in the yaml_test_runner
package so that they can be invoked with cargo test.

The api_generator package is now exposed as a library
to yaml_test_runner package, whilst also still compiling it as a
runnable binary to generate the client. This allows the generator modules
to be used in the yaml_test_runner where they are used to generate tests,
by inspecting the ASTs produced by the api_generator.

The typical flow to generate and run tests is

1. Define ELASTICSEARCH_VERSION environment variable

  export ELASTICSEARCH_VERSION=elasticsearch-oss:7.7.0-SNAPSHOT

  This is used to determine which test suite to generate, based on whether
  the version uses the default distribution (xpack suite), or the oss
  distribution (oss suite).

2. Run yaml_test_runner

  cargo run -p yaml_test_runner -- \
      --branch <elasticsearch branch> \
      --token <token> \
      --path "<path to rest specs>"

  where
  - --branch is the elasticsearch branch to target to download yaml tests
  - --token is a GitHub access token to use to download the yaml tests using the content API
  - --path is the path to download rest specs to

3. Run the generated tests

  cargo test -p yaml_test_runner -- --test-threads=1

  The tests must be run synchronously, so a single thread must be used.

Other pertinent changes in this commit:

* impl From<> for TrackTotalHits for i64 and bool
* Define expand_wildcards param as a collection
* Handle url serialization of enum collections

  Updates serialize_coll_qs to handle serializing
  a collection of types that implement Serialize to a comma separated
  query string value. This is only needed
  for enums and strings, but (mis)uses serde_json serialize to
  achieve this, since serde_urlencoded does not support such a format.
  We could probably implement a serde Serializer for this later on.

* Credentials set before setting any other headers in transport
  to allow headers to overwrite values.
* Update TypeKind::None to TypeKind::Unknown(String)
  Better handle new types when introduced

Closes #19

(cherry picked from commit 27b8ac9e4990b96c85096b5bd5a580997bbd4ee1)