Add CI jobs and scripts (#108)
This commit adds the CI jobs and build scripts to run the yaml
tests for the client in CI.

The Dockerfile is updated to run the yaml tests. A specific rust nightly
version is defined that defines the rust toolchain used; current rust
nightly (2020-06-12) is broken when attempting to build this project
so a known good nightly version is used, which can be overridden with
RUST_TOOLCHAIN environment variable.

Remove PowerShell scripts as shell scripts can be called from
WSL2 with Docker on Windows.