Download yaml tests from tarball
This commit updates how yaml tests are
downloaded by streaming the tarball for
a specific branch, and extracting yaml tests
from the tarball. Not only is this much faster
than downloading separate files, it removes
the need for a GitHub access token as
anonymous access limits are likely sufficient.

Don't download rest specs when running
yaml_test_runner as the checked in rest specs
will be those that have been used to generate
the client, and therefor should be the ones
used when generating tests using the client.