Add Exception to model Elasticsearch exceptions (#118)
This commit adds an Exception struct that models
exceptions returned by Elasticsearch in the event of an error.

An exception() function is exposed on Response to deserialize the
response body into Exception when the status code is
in the 400-599 range.

Consideration was made to exposing this on
the Error struct, performing the deserialization inside
`error_for_status_code`, but this introduces some awkwardness into
the API:

- `error_for_status_code` would ideally be async because the .json()
  call is async
- .json() consumes the response, which would not be possible for
  `error_for_status_code_ref` which has a shared reference to Response.

Consideration was made to combining `Error` and `Cause` into one type, since
they both map to `ElasticsearchException` on the server side. It looks like
certain properties are only ever exposed on the top level exception however,
which is deserialized to `Error` e.g.
root_cause, header and failed_shards (which is not currently mapped in Error).
In light of this, decided to keep them separate for now.

Additional details about an exception are collected in a map and exposed
through additional_details function.

Closes #112

(cherry picked from commit b620d2207a74a6e414938880e9eb4cc717297aae)