Create platform specific run-elasticsearch task (#159)
This commit splits out the command and script/script runner elements of the
run-elasticsearch task to fix a bug when running on Windows where
the task on linux and macOS uses a command
whilst windows uses a script to run via WSL2. The Windows specific
task ends up with multiple actions (command and script) which
cargo make errors on.