Add bytes method to Response struct (#164)
* Use a conventional struct instead of a Tuple struct

The Response struct contains a reqwest::Request and a Method fields, and
accessing them by position may be a bit confusing. Giving an actual name
to the fields brings us the benefit of easily understanding the methods
we are reading, without the need to go to the top to remember which
field we are referring to.

Reference: https://doc.rust-lang.org/1.9.0/book/structs.html#tuple-structs

* Expose `bytes` to the Response object

This patch exposes the `Response::bytes`, which is basically a wrapper
on reqwest::Response::bytes().

* Add additional note on vm.max_map_count settings

Closes #160

(cherry picked from commit 96189a5b2000c0b32b414034275528b77567ee76)