Add update-version task (#172)
This commit adds an update version task to cargo make.

The task is implemented as a rust script to work cross-platform
and to allow retrieving the current version from Cargo.toml
package version, for replacement in docs. It assumes that
the version to be updated across all docs is the same.